Rename the i686 module to x86
This module is used for both i686 and x86-64.
